changed individual location/episode/character page from server side render to static page.
